move types and update imports and exports
