+++ packages/web-components/src/accordion-item/index.ts
@@ -1,6 +1,9 @@
+export {
+  baseName as AccordionItemBaseName,
+  definition as AccordionItemDefinition,
+  isAccordionItem,
+} from './accordion-item.definition.js';
 export { AccordionItem, BaseAccordionItem } from './accordion-item.js';
-export type { AccordionItemOptions } from './accordion-item.js';
-export { AccordionItemSize, AccordionItemMarkerPosition } from './accordion-item.options.js';
+export { AccordionItemMarkerPosition, AccordionItemSize, type AccordionItemOptions } from './accordion-item.options.js';
 export { styles as accordionItemStyles } from './accordion-item.styles.js';
-export { definition as accordionItemDefinition } from './accordion-item.definition.js';
 export { template as accordionItemTemplate } from './accordion-item.template.js';
